add transport logic

+++ goshop-facade-goods/src/main/java/org/goshop/goods/pojo/GoodsClassStaple.java
@@ -3,27 +3,27 @@ package org.goshop.goods.pojo;
 import java.io.Serializable;
 
 public class GoodsClassStaple implements Serializable {
-    private Integer stapleId;
+    private Long stapleId;
 
     private String stapleName;
 
-    private Integer gcId1;
+    private Long gcId1;
 
-    private Integer gcId2;
+    private Long gcId2;
 
-    private Integer gcId3;
+    private Long gcId3;
 
-    private Integer typeId;
+    private Long typeId;
 
     private Long memberId;
 
     private Integer counter;
 
-    public Integer getStapleId() {
+    public Long getStapleId() {
         return stapleId;
     }
 
-    public void setStapleId(Integer stapleId) {
+    public void setStapleId(Long stapleId) {
         this.stapleId = stapleId;
     }
 
@@ -35,35 +35,35 @@ public class GoodsClassStaple implements Serializable {
         this.stapleName = stapleName == null ? null : stapleName.trim();
     }
 
-    public Integer getGcId1() {
+    public Long getGcId1() {
         return gcId1;
     }
 
-    public void setGcId1(Integer gcId1) {
+    public void setGcId1(Long gcId1) {
         this.gcId1 = gcId1;
     }
 
-    public Integer getGcId2() {
+    public Long getGcId2() {
         return gcId2;
     }
 
-    public void setGcId2(Integer gcId2) {
+    public void setGcId2(Long gcId2) {
         this.gcId2 = gcId2;
     }
 
-    public Integer getGcId3() {
+    public Long getGcId3() {
         return gcId3;
     }
 
-    public void setGcId3(Integer gcId3) {
+    public void setGcId3(Long gcId3) {
         this.gcId3 = gcId3;
     }
 
-    public Integer getTypeId() {
+    public Long getTypeId() {
         return typeId;
     }
 
-    public void setTypeId(Integer typeId) {
+    public void setTypeId(Long typeId) {
         this.typeId = typeId;
     }
